+++ ReleaseNotes.md
@@ -18,6 +18,7 @@ Current Support
 * The pipeline state will be displayed at each command, showing the data contained in each member of the pipeline createinfo struct, as well as dynamic state.
 * Simple disassembly/reflection of SPIR-V to determine which descriptors to read for read-only resources and uniform buffers. The uniform buffers will be listed separately and the member variables filled out.
 * Simple display of most 2D textures with mips in the texture viewer.
+* Drawcall highlight and wireframe overlays
 * Threading should be pretty efficient - no heavy locks on common paths (outside of creation/deletion)
 
 Known Issues
@@ -42,7 +43,7 @@ On replay:
 
 * Only 2D non-array non-integer textures can currently be displayed.
 * Auto texture range-fit or histogram display is not implemented.
-* Debug overlays aren't implemented.
+* Debug overlays other than drawcall highlight and wireframe aren't implemented.
 * Saving textures is not supported.
 * Queue-level API events are not properly listed. API calls between draw-type vkCmd... are listed.
 * Meshes are not rendered as a preview.